Move Helper class out of Tests dir to avoid be executed
